fix(trainer): import correct sb[3] BaseCallback
It seems sb3 is being used for training the battler agent, and sb for
the drafting agents.  However, the `BaseCallback` import statement
imports from sb3.  For the drafting agents, the PPO2 model is passed
this sb3 `BaseCallback` object, which errors out when it is called
because it expects a different interface to models (it expects a logger
object on the model).

Create two different `TrainingCallback` classes,
`BattleTrainingCallback` and `DraftTrainingCallback`, which inherit from
the appropriate `BaseCallback`, and add logic in the rest of the file to
initialize callbacks from the correct class based on the training task.

There is already logic handling something similar for `CallbackList`
vs. `CallbackList3`.  Consolidate the logic for both of these cases into
a single function to reduce code redundancy.