fix(trainer): correct evaluator stats calculation
The evaluator updates statistics for all parallel environments at once
before checking whether the stop condition has been met.  This means
that if the stop condition (number of episodes) is not a multiple of the
concurrency (number of parallel environments), you may end up with some
extra episodes included in the stats.

The previous version of the code handled this by truncating the stats,
but the order of the stats meant this was truncating the last few
episodes of the last environment, instead of truncating the last extra
episode of each environment.

So, assuming concurrency of 3 and 100 evaluation episodes, for the
previous code, at the end of the while loop:

    >>> [len(rewards[:-1]) for rewards in episode_rewards]
    [34, 34, 34]

(the ``[:-1]`` is to drop the final, incomplete, accumulating stat).

Then they are combined into a single list, which ends up being
equivalent to:

    >>> (episode_rewards[0][:-1]
    ... + episode_rewards[1][:-1]
    ... + episode_rewards[2][:-1])

And finally the extra values (2 of them, 34*3 - 100) are truncated.
Since ``episdoe_rewards[2][:-1]`` is at the end of the list, the last
two stats from environment 2 are truncated.  While it is unlikely to be
significant unless the concurrency is large, what should happen is
instead that the last stat from ``episode_rewards[1][:-1]`` and the last
state from ``episode_rewards[2][:-1]`` each be dropped.

To get this result instead, this fix stops appending new stats once the
stop condition is reached, even if it is in the middle of a loop over
the environments.  So at the end of the while loop:

    >>> [len(rewards[:-1]) for rewards in episode_rewards]
    [34, 33, 33]

and truncation is no longer necessary.

Could have also done this by transposing the episode rewards matrix
before concatenating it and truncating the result.