fix(trainer): prevent callback from discarding obs
The previous approach to terminating training based on number of
completed episodes was not quite correct.  It used the `on_step`
callback, which is run after each environment step when generating
rollouts.  To determine when to stop, the callback checks whether
``env.episodes``  exceeds the next switch count, and if so, returns
`False` to terminate the training.

However, this doesn't get the desired result.  It immediately terminates
the current batch of rollouts, and ends the learning invocation *without
training the model on this final batch*.  This means the whole batch of
observations for that rollout are discarded.

Depending on the number of parallel environments and `n_steps`
parameter, this means that a possibly large number of episodes for each
call to `learn` are never trained on.  At the very least, the last
episode to complete, which triggered the cancel, doesn't get trained on,
even though it should be.

Unfortunately, there is no elegant way to get more correct behavior,
because the `stable-baselines` models use timesteps rather than
completed episodes for determining how long to learn.  Even worse, the
`on_step` callback doesn't run until *after* stepping the environments,
so there's no way to use that callback to terminate training without
leaving the environments in an improper state.

This fix uses an admittedly hacky workaround.

Instead of the `on_step` callback, it uses the `on_rollout_start` and
`on_rollout_end` callbacks.  The check for the training end condition is
performed during `on_rollout_start`, after the rollout from the previous
loop was used for training, but before another rollout is generated.

If the condition is met, the callback manually sets the
`continue_training` flag on the runner to `False`.  It then replaces the
`run` method on the runner with a dummy which simply returns an empty
dummy vector of results (the same behavior as the real `run` when a call
to `on_step` returns `False`).  This allows us to skip generating
rollouts, and seems to be the only practical way to do so.

Finally, after skipping the rollouts, the `on_rollout_end` callback
restores the original `run` method to the runner.

Since `continue_training` on the runner is still `False` at this point,
training then terminates gracefully, as desired.