feat(engine): add drawn_cards to state
Add a `drawn_cards` list to the player state.  Append to this list each
time the player draws, and reset the list at the beginning of each
action.  This allows environments to observe draw events as they happen,
for reporting to the learning agent.

This is part of setting up for using evaluation battle draw order to
change how rewards are allocated when training draft agents.