fix: correct and simplify seed management
The previous approach was to increment the environment/state/agent seed
every time the environment is reset, so each episode starts with a new
seed.  It also set the initial seed for each environment so that the
seeds for the environments didn't overlap.

However, the `stable-baselines` training logic very helpfully overwrites
all of the environment seeds that took such careful work to set up.
What `stable-baselines` does is set the seed of the first environment to
the seed provided when initializing the model.  It then increments the
seed by one for each environment.

So instead of (assuming 4 training steps and four environments):
[0, 4, 8, 12]

We get:
[0, 1, 2, 3]

Then, when the environment increments the seeds upon reset, instead of
this:
[1, 5, 9, 13]

We get:
[1, 2, 3, 4]

And end up with three repeated seeds.

To work better with `stable-baselines`, this patch removes all manual
setting of environment seeds and lets the model handle that.  It removes
the seed incrementing on environment reset, both to prevent repeat
seeds, and because we don't really need to set the seeds again anyway.
Finally, it applies similar logic/setup to the `Evaluator`