feat(trainer): determine reward using draw order
During the `on_step` callback, if it is the end of an episode, use the
`drafted` and `drawn` entries in the `info` dictionaries to determine
the order cards were drawn in the evaluation battle, and which ones
weren't drawn.

Use this information to update the `mb_effective_steps` local variable
in the `Runner`.  Map each draft step to the step corresponding to when
the chosen card was drawn in battle.  Count battle steps back from the
terminal step.  Prepend all undrawn cards, and update the `mb_gae_maskS`
variable to indicate they should be skipped during the GAE calculation.

The effect of all if this is to ensure that the reward is allocated only
to cards which were actually drawn in battle, and to allocate the most
reward to the cards drawn closest to victory or loss.  This more
accurately matches the intent of the PPO algorithm and leads to more
reliable training and final performance.